# toteco_openapi.yaml
openapi: 3.0.2
info:
  title: TOTECO API
  description: API to manage publications about products (coffee, tea or toast) like a social media. The idea is to create a community that publishes photos of their toasts with coffee or tea adding the price and being able to rate the products and the establishments.
  version: "1.0"
  contact:
    email: a25540@svalero.com
servers:
  - url: http://localhost:8081
...
paths:
  /products:
    get:
      tags:
        - Products
      summary: Get all the products
      description: Get all the products in the DDBB.
      operationId: getAllProducts
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
          
    post:
      security:
        - bearerAuth: []
      tags:
        - Products
      summary: Post a new product
      description: Save a new product in the DDBB.
      operationId: saveProduct
      requestBody:
        description: product to save
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
      responses:
        201:
          description: Created
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        400:
          description: Bad request exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        404:
          description: Not found exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
                
    delete:
      security:
        - bearerAuth: []
      tags:
        - Products
      summary: Delete all the products
      description: Delete all the products from the DDBB.
      operationId: deleteAllProducts
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: string
                example: All products deleted.
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get a product
      description: Get an existing product by the id
      operationId: getProductById
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
      
    put:
      security:
        - bearerAuth: []
      tags:
        - Products
      summary: Update product
      description: Update an existing product
      operationId: updateProduct
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        description: product to save
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
      
    delete:
      security:
        - bearerAuth: []
      tags:
        - Products
      summary: Delete all the products
      description: Delete all the products in the DDBB
      operationId: deleteProductById
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: string
                example: Product deleted.
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /products/date/{date}:
    get:
      tags:
        - Products
      summary: Get products by date
      description: Get all the products which have this creation date. The date format is timestamp in miliseconds.
      operationId: getProductsByDate
      parameters:
        - $ref: '#/components/parameters/date'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /products/date/between:
    get:
      tags:
        - Products
      summary: Get products between dates
      description: Get all the products which have the creation date between theese. The dates format is timestamp in miliseconds
      operationId: getProductsByDateBetween
      parameters:
        - $ref: '#/components/parameters/minDate'
        - $ref: '#/components/parameters/maxDate'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /products/inMenu/{inMenu}:
    get:
      tags:
        - Products
      summary: Get products which belongs (or not) to menus
      description: Get all the products which are part of any menu or get all the products that are not par of any menu.
      operationId: getProductsByInMenu
      parameters:
        - $ref: '#/components/parameters/inMenu'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /products/{id}/price/{price}:
    patch:
      security:
        - bearerAuth: []
      tags:
        - Products
      summary: Update product price
      description: Update the price of the indicated product.
      operationId: updateProductsPrice
      parameters:
        - $ref: '#/components/parameters/id'
        - $ref: '#/components/parameters/price'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        404:
          description: Not found exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /products/price/{price}:
    get:
      tags:
        - Products
      summary: Get all products by price
      description: Get all the products which have this same price.
      operationId: getProductsByPrice
      parameters:
        - $ref: '#/components/parameters/price'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /products/price/between:
    get:
      tags:
        - Products
      summary: Get products by price between
      description: Get all the products which prices are between theese two.
      operationId: getProductsByPriceBetween
      parameters:
        - $ref: '#/components/parameters/minPrice'
        - $ref: '#/components/parameters/maxPrice'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
...
components:
  schemas:
    Product:
      properties:
        id: 
          type: integer
          format: int64
          example: 0
        type:
          $ref: '#/components/schemas/Type'
        date:
          type: string
          format: dd-MM-yyyy
          example: 04-03-2022
        price:
          type: number
          example: 3.40
        score:
          type: number
          example: 4.3
        menu:
          $ref: '#/components/schemas/Menu'
        publication:
          $ref: '#/components/schemas/Publication'          
    ProductDTO:
      properties:
        inMenu:
          type: boolean
          example: false
        price:
          type: number
          example: 3.40
        score:
          type: number
          example: 4.3
        typeId:
          type: integer
          format: int64
          example: 0
        menuId:
          type: integer
          format: int64
          example: 0
        publicationId:
          type: integer
          format: int64
          example: 0
...
    Exception:
      properties:
        code: 
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
        message:
          type: string
  parameters:
    id:
      name: id
      in: path
      description: id of the element
      required: true
      schema:
        type: integer
        format: int64
        example: 0
    date:
      name: date
      in: path
      description: creation date
      required: true
      schema:
        type: integer
        example: 1640604419000
    minDate:
      name: minDate
      in: query
      description: date from in timestamp miliseconds
      required: true
      schema:
        type: integer
        example: 1640604419000
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT